move dataclasses to types
